% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_binary}
\alias{check_binary}
\title{check_binary}
\usage{
check_binary(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A logical value
}
\description{
Check if the outcome variable is 0/1 binary. Return 1 if true, 0
if false
}
