% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_entry_errors}
\alias{check_entry_errors}
\title{Check arguments to drcmd for entry errors}
\usage{
check_entry_errors(
  Y,
  A,
  X,
  W,
  eem_ind,
  Rprobs,
  k,
  cutoff = 0.025,
  cv_folds = 5,
  tml = FALSE,
  quiet = TRUE,
  parallel = FALSE,
  att = FALSE,
  atc = FALSE
)
}
\arguments{
\item{Y}{A vector or data frame containing outcome values}

\item{A}{A vector or data frame  containing treatment variable values}

\item{X}{A data frame containing covariate values}

\item{W}{A data frame containing proxy variable values}
}
\description{
Checks if the arguments to main function drcmd are correctly specified.
Returns TRUE if all checks are passed, throws an error with message otherwise.
}
\examples{
\dontrun{
n <- 200
X <- data.frame(X1 = rnorm(n))
W <- data.frame(W1 = rnorm(n))
A <- rbinom(n, 1, 0.5)
Y <- rnorm(n)
check_entry_errors(Y, A, X, W, eem_ind = FALSE, Rprobs = NA, k = 1)
}
}
\keyword{internal}
