% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_r_ind}
\alias{check_r_ind}
\title{check_r_ind}
\usage{
check_r_ind(data, Y, A, X, W, R)
}
\arguments{
\item{data}{A data frame}

\item{Y}{A character string containing outcome variable name}

\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing covariate variable names}

\item{W}{A character vector containing weight variable names}

\item{R}{A character string containing randomization variable name}
}
\value{
A logical value
}
\description{
Check if the missingness indicator R is defined correctly
}
