% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_crossfit_nuis}
\alias{clean_crossfit_nuis}
\title{Clean nuisance function output from crossfit procedure}
\usage{
clean_crossfit_nuis(results)
}
\arguments{
\item{results}{Nuisance functions output from drcmd_est}
}
\value{
A dataframe of averaged nuisance estimates across all folds
}
\description{
Transforms nuisance output across folds into a dataframe of avgerge
prediction at each point for each nuisance function
}
