% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_learners}
\alias{clean_learners}
\title{Clean SuperLearner libraries}
\usage{
clean_learners(
  default_learners,
  m_learners,
  g_learners,
  r_learners,
  po_learners
)
}
\arguments{
\item{default_learners}{Either null, or a character vector containing SuperLearner
libraries to use for estimating all nuisance functions. User can alternatively
specify libraries for each nuisance function for added flexibility}

\item{m_sl_learners}{Either null, or a character vector containing SuperLearner
libraries to be used for the outcome regression}

\item{g_sl_learners}{Either null, or a character vector containing SuperLearner
libraries to be used for the propensity scores}

\item{r_sl_learners}{Either null, or a character vector containing SuperLearner
libraries to be used for the missingness indicator regression}

\item{po_sl_learners}{Either null, or a character vector containing SuperLearner
libraries to be used for the pseudo outcome regression}
}
\value{
A list of
}
\description{
Internal function for setting SuperLearner libraries before they
are passed into estimation procedures. Default libraries from the sl_learners
argument are used to fill in missing libraries for any nuisance function with
libraries left unspecified
}
