% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{drcmd_est}
\alias{drcmd_est}
\title{Obtain doubly-robust counterfactual mean estimates through cross-fitting}
\usage{
drcmd_est(
  Y,
  A,
  X,
  Z,
  R,
  m_learners,
  g_learners,
  r_learners,
  po_learners,
  eem_ind,
  tml,
  Rprobs,
  k,
  cutoff,
  y_bin,
  yscaled,
  cv_folds = 5,
  parallel = FALSE,
  quiet = TRUE,
  att = FALSE,
  atc = FALSE
)
}
\arguments{
\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{Z}{Dataframe containing all variables that are never subject to missingness}

\item{R}{Binary missingness indicator}

\item{m_learners}{A character vector containing learners to be used for the
outcome regression. User can specify 'hal' or a vector of SuperLearner libraries}

\item{g_learners}{A character vector containing learners to be used for the
propensity score. User can specify 'hal' or a vector of SuperLearner libraries}

\item{r_learners}{A character vector containing learners to be used for the
missingness indicator regression. User can specify 'hal' or a vector of
SuperLearner libraries}

\item{po_learners}{A character vector containing learners to be used for the
pseudo-outcome regression. User can specify 'hal' or a vector of SuperLearner libraries}

\item{eem_ind}{A logical indicating whether to use empirical efficiency maximization}

\item{Rprobs}{A vector of probabilities for R}

\item{k}{A numeric indicating the number of folds for cross-fitting}

\item{data}{A data frame}
}
\value{
A list of point estimates and standard errors for all estimands considered
}
\description{
Outer function for estimating counterfactual means through cross-fitting.
Calls \code{drcmd_est_fold} to obtain estimates for each cross-fitting fold
}
\examples{
\dontrun{
n <- 500
X <- data.frame(X1 = rnorm(n))
A <- rbinom(n, 1, plogis(X$X1))
Y <- A + X$X1 + rnorm(n)
R <- rbinom(n, 1, plogis(X$X1))
Y[R == 0] <- 0; A[R == 0] <- 0
Z <- X
res <- drcmd_est(Y, A, X, Z, R,
                 m_learners = "SL.glm", g_learners = "SL.glm",
                 r_learners = "SL.glm", po_learners = "SL.glm",
                 eem_ind = FALSE, tml = FALSE,
                 Rprobs = NA, k = 1, cutoff = 0.025,
                 y_bin = FALSE, yscaled = FALSE)
}
}
