% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{drcmd_est_fold}
\alias{drcmd_est_fold}
\title{drcmd_est_fold}
\usage{
drcmd_est_fold(splits, Y, A, X, Z, R, hal_ind, sl.lib, eem_ind, Rprobs)
}
\arguments{
\item{splits}{A list of train/test splits}

\item{Y}{A character string}

\item{A}{A character string}

\item{X}{A character vector}

\item{Z}{A character vector}

\item{R}{A character vector}

\item{hal_ind}{A logical indicating whether to use highly adaptive lasso}

\item{sl.lib}{A character string indicating the superlearner library to use}

\item{eem_ind}{A logical indicating whether to use ensemble of estimators}

\item{Rprobs}{A vector of probabilities for R}

\item{data}{A data frame}

\item{k}{A numeric indicating the number of folds for cross-fitting}
}
\value{
A list of results
}
\description{
Outer function for obtaining point estimates for single fold
}
