% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_g}
\alias{est_g}
\title{est_g}
\usage{
est_g(idx, A, X, R, kappa_hat, hal_ind, sl.lib)
}
\arguments{
\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing covariate variable names}

\item{R}{A character string containing randomization variable name}

\item{hal_ind}{A logical value indicating whether to estimate via HAL}

\item{sl.lib}{A character vector containing the names of the superlearner algorithms}

\item{data}{A data frame}

\item{Y}{A character string containing outcome variable name}

\item{W}{A character vector containing weight variable names}
}
\value{
A list containing the estimate of E\link{Y|A=a,X,W}
}
\description{
Function for obtaining estimate of E\link{Y|A=a,X,W}
}
