% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_m_a}
\alias{est_m_a}
\title{est_m_a}
\usage{
est_m_a(idx, Y, A, X, R, kappa_hat, hal_ind, sl.lib)
}
\arguments{
\item{Y}{A character string containing outcome variable name}

\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing the names of the variables in X}

\item{R}{A character string containing randomization variable name}

\item{kappa_hat}{A numeric vector containing the fitted values of kappa}

\item{hal_ind}{A logical value indicating whether to estimate via HAL}

\item{sl.lib}{A character vector containing the names of the superlearner algorithms}

\item{data}{A data frame}
}
\value{
A list containing the estimate of E\link{Y|A=a,X}
}
\description{
Function for obtaining estimate of E\link{Y|A=a,X}
}
