% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{est_psi}
\alias{est_psi}
\title{est_psi}
\usage{
est_psi(idx, R, Z, kappa_hat, phi_hat, varphi_hat)
}
\arguments{
\item{R}{A character string containing randomization variable name}

\item{Z}{A character vector containing the names of the variables in Z}

\item{kappa_hat}{A numeric vector containing the fitted values of kappa}

\item{phi_hat}{A list containing the estimate of E\link{phi|Z}}

\item{varphi_hat}{A list containing the estimate of E\link{phi|Z}}

\item{data}{A data frame}
}
\value{
A numeric value containing the estimate of E\link{psi}
}
\description{
Function for obtaining estimate of the full data EIC, psi
}
