% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{est_varphi}
\alias{est_varphi}
\title{est_varphi}
\usage{
est_varphi(idx, R, Z, phi_1_hat, phi_0_hat, hal_ind, sl.lib)
}
\arguments{
\item{R}{A character string containing randomization variable name}

\item{phi_1_hat}{A numeric vector containing the fitted values of phi under A=1}

\item{phi_0_hat}{A numeric vector containing the fitted values of phi under A=0}

\item{hal_ind}{A logical value indicating whether to estimate via HAL}

\item{sl.lib}{A character vector containing the names of the superlearner algorithms}

\item{data}{A data frame}
}
\value{
A list containing the estimate of E\link{phi|Z}
}
\description{
Function for obtaining estimate of E\link{phi|Z}
}
