% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_varphi_main}
\alias{est_varphi_main}
\title{Perform pseudo-outcome regression}
\usage{
est_varphi_main(
  idx,
  R,
  Z,
  phi_1_hat,
  phi_0_hat,
  kappa_hat,
  eem_ind,
  po_learners
)
}
\arguments{
\item{idx}{Indices to carry out estimation over}

\item{R}{Binary missingness indicator, where 0 indicates missing data}

\item{Z}{Dataframe containing all non-missing variables}

\item{eem_ind}{A logical value indicating whether to estimate via EEM}

\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{phi_hat}{A list containing the estimate of phi}
}
\value{
A list containing the estimate of E\link{phi_a|Z} for a=0 and a=1
}
\description{
Function for obtaining estimate of E\link{phi_a|Z} through pseudo-outcome
regression. Calls inner function to perform estimation, depending on whether
user wishes to perform empirical efficiency maximization or not
}
