% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_missing_pattern}
\alias{find_missing_pattern}
\title{find_missing_pattern}
\usage{
find_missing_pattern(Y, A, X, W, min_complete = 10L)
}
\arguments{
\item{Y}{A vector or data frame containing outcome values}

\item{A}{A vector or data frame  containing treatment variable values}

\item{X}{A data frame containing covariate values}

\item{W}{A data frame containing proxy variable values}

\item{min_complete}{Number of complete cases below which to issue a warning}
}
\value{
A character string containing the missing pattern
}
\description{
Find the missing pattern in the data
}
\examples{
\dontrun{
n <- 200
X <- data.frame(X1 = rnorm(n))
A <- rbinom(n, 1, 0.5)
Y <- rnorm(n)
R <- rbinom(n, 1, 0.7)
Y[R == 0] <- NA
W <- X[, 0]
result <- find_missing_pattern(Y, A, X, W)
result$Z  # variables without missingness
result$U  # variables with missingness
result$R  # complete case indicator
}
}
\keyword{internal}
