% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{get_nuisance_ests}
\alias{get_nuisance_ests}
\title{Obtain nuisance function estimates}
\usage{
get_nuisance_ests(
  idx,
  Y,
  A,
  X,
  Z,
  R,
  m_learners,
  g_learners,
  r_learners,
  Rprobs,
  cutoff,
  cv_folds = 5,
  quiet = TRUE
)
}
\arguments{
\item{idx}{Indices to carry out estimation over}

\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{Z}{Dataframe containing all non-missing variables}

\item{R}{Binary missingness indicator, where 0 indicates missing data}

\item{m_learners}{A character vector containing learners to be used for the
outcome regression. User can specify 'hal' or a vector of SuperLearner libraries}

\item{g_learners}{A character vector containing learners to be used for the
propensity score. User can specify 'hal' or a vector of SuperLearner libraries}

\item{r_learners}{A character vector containing learners to be used for the
missingness indicator regression. User can specify 'hal' or a vector of
SuperLearner libraries}

\item{Rprobs}{A vector of probabilities for R}

\item{eem_ind}{A logical indicating whether to use empirical efficiency maximization}
}
\value{
A list of nuisance estimates
}
\description{
Function for obtaining estimates of all relevant nuisance functions
}
\examples{
\dontrun{
n <- 500
X <- data.frame(X1 = rnorm(n))
A <- rbinom(n, 1, plogis(X$X1))
Y <- A + X$X1 + rnorm(n)
R <- rbinom(n, 1, plogis(X$X1))
Y[R == 0] <- 0; A[R == 0] <- 0
Z <- X
nuis <- get_nuisance_ests(1:n, Y, A, X, Z, R,
                          m_learners = "SL.glm",
                          g_learners = "SL.glm",
                          r_learners = "SL.glm",
                          Rprobs = NA, cutoff = 0.025)
}
}
