% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{get_phi_hat}
\alias{get_phi_hat}
\title{Contruct full data EIF estimate}
\usage{
get_phi_hat(Y, A, X, R, Z, g_hat, m_a_hat, kappa_hat, idx = seq_along(Y))
}
\arguments{
\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{R}{Binary missingness indicator}

\item{Z}{Dataframe containing all variables that are never subject to missingness}

\item{g_hat}{Propensity score predictions}

\item{m_a_hat}{List of outcome predictions under A=0/1}

\item{kappa_hat}{Missingness probabilities}

\item{idx}{Indices used to compute the plug-in estimates. Defaults to all
observations; during cross-fitting these are the training-fold indices.}
}
\description{
Function for constructing estimate of the full data EIF, given
propensity score and outcome model fits
}
\examples{
\dontrun{
n <- 500
X <- data.frame(X1 = rnorm(n))
A <- rbinom(n, 1, plogis(X$X1))
Y <- A + X$X1 + rnorm(n)
R <- rep(1, n)
Z <- X
g_hat <- plogis(X$X1)
m_a_hat <- list(m_1_hat = 1 + X$X1, m_0_hat = X$X1)
kappa_hat <- rep(1, n)
phi <- get_phi_hat(Y, A, X, R, Z, g_hat, m_a_hat, kappa_hat)
}

}
\keyword{internal}
