% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.drcmd}
\alias{plot.drcmd}
\title{Plot results from drcmd object}
\usage{
\method{plot}{drcmd}(x, type = "All", ...)
}
\arguments{
\item{x}{An object of class drcmd}

\item{type}{Character denoting type of plot to generate. Must be one of 'All', 'PO',
'IC', 'g_hat', 'r_hat'}

\item{...}{Additional arguments (unused)}
}
\value{
No return value. Called for plotting results from drcmd object
}
\description{
S3 method for plotting results from drcmd object. Plots are
available for the following: (1) pseudo-outcome regression fit, (2) influence
curve distribution, (3) treatment propensity score distribution, and (4)
complete-case propensity score distribution. When type='All' (the default),
user can view all four plots in succession interactively
}
\examples{
set.seed(1)
n <- 200
X <- rnorm(n)
A <- rbinom(n, 1, plogis(X))
Y <- rnorm(n) + A + X
R_ind <- rbinom(n, 1, plogis(X))
Y[R_ind == 0] <- NA
covariates <- data.frame(X = X)

fit <- drcmd(Y, A, covariates, default_learners = "SL.glm", k = 1)
plot(fit, type = "PO")
}
