% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{predict.SL.hal9001}
\alias{predict.SL.hal9001}
\title{Prediction wrapper for the highly-adaptive lasso}
\usage{
\method{predict}{SL.hal9001}(object, newdata, ...)
}
\arguments{
\item{...}{Additional arguments to pass to the HAL function}

\item{Y}{Outcome variable}

\item{X}{Matrix of covariates}

\item{newX}{Matrix of covariates for prediction}

\item{family}{Family of the outcome variable}

\item{obsWeights}{Observation weights}
}
\value{
Predicted values on new data
}
\description{
Prediction wrapper for the highly-adaptive lasso (HAL) algorithm
implemented in the hal9001 package
}
