% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.drcmd}
\alias{print.drcmd}
\title{Print drcmd object}
\usage{
\method{print}{drcmd}(x, ...)
}
\arguments{
\item{x}{An object of class drcmd}

\item{...}{Additional arguments (unused)}
}
\value{
No return value. Called for printing a concise results summary
}
\description{
S3 method for printing drcmd objects. Provides concise summary
of results, and prints values of optional arguments. Use summary() function for
more detailed summary of results.
}
\examples{
set.seed(1)
n <- 200
X <- rnorm(n)
A <- rbinom(n, 1, plogis(X))
Y <- rnorm(n) + A + X
R_ind <- rbinom(n, 1, plogis(X))
Y[R_ind == 0] <- NA
covariates <- data.frame(X = X)

fit <- drcmd(Y, A, covariates, default_learners = "SL.glm", k = 1)
print(fit)
}
