% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{SL.hal9001}
\alias{SL.hal9001}
\title{SuperLearner wrapper for the highly-adaptive lasso}
\usage{
SL.hal9001(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{Outcome variable}

\item{X}{Matrix of covariates}

\item{newX}{Matrix of covariates for prediction}

\item{family}{Family of the outcome variable}

\item{obsWeights}{Observation weights}

\item{...}{Additional arguments to pass to the HAL function}
}
\value{
A list containing the prediction and the fitted model
}
\description{
Wrapper for the highly-adaptive lasso (HAL) algorithm
implemented in the hal9001 package
}
\examples{
\dontrun{
n <- 200
X <- data.frame(X1 = rnorm(n))
Y <- X$X1 + rnorm(n)
fit <- SL.hal9001(Y, X, newX = X, family = gaussian(),
                  obsWeights = rep(1, n))
head(fit$pred)
}
}
