% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.drcmd}
\alias{summary.drcmd}
\title{Summarize results from drcmd}
\usage{
\method{summary}{drcmd}(object, detail = FALSE, ...)
}
\arguments{
\item{object}{An object of class drcmd}

\item{detail}{Logical. If TRUE, print out values of user-supplied arguments}

\item{...}{Additional arguments (unused)}
}
\value{
No return value. Called for printing a detailed results summary
}
\description{
S3 method for summarizng drcmd results. Provides detailed summary
of estimation output, while also providing information on missingness
mechanism. Can optionally print out values of user-supplied arguments by setting
detail=TRUE
}
\examples{
set.seed(1)
n <- 200
X <- rnorm(n)
A <- rbinom(n, 1, plogis(X))
Y <- rnorm(n) + A + X
R_ind <- rbinom(n, 1, plogis(X))
Y[R_ind == 0] <- NA
covariates <- data.frame(X = X)

fit <- drcmd(Y, A, covariates, default_learners = "SL.glm", k = 1)
summary(fit)
}
