% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_r}
\alias{truncate_r}
\title{Truncate complete case propensity scores}
\usage{
truncate_r(x, c = 0.01)
}
\arguments{
\item{x}{A vector of complete case propensity scores}
}
\value{
A vector of complete propensity scores truncated to interval \link{c, 1-c}
}
\description{
Truncate propensity scores to interval \link{c, 1-c}
}
