% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{drcmd}
\alias{drcmd}
\title{Doubly-robust causal inference with missing data}
\usage{
drcmd(
  Y,
  A,
  X,
  W = NA,
  R = NA,
  hal_ind = TRUE,
  sl.lib = NA,
  eem_ind = FALSE,
  Rprobs = NA,
  k = 1,
  nboot = 0
)
}
\arguments{
\item{Y}{A character string containing outcome variable name}

\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing covariate names}

\item{W}{A character vector containing variable names solely used for imputation}

\item{R}{(optional) A character string specifying the missingness indicator, where 0 indicates missing data. If not specified, the function will identify the missingness indicator}

\item{hal_ind}{A logical indicating whether to use highly adaptive lasso}

\item{sl.lib}{A character string indicating the superlearner library to use}

\item{eem_ind}{A logical indicating whether to use ensemble of estimators}

\item{Rprobs}{A vector of probabilities for R}

\item{k}{A numeric indicating the number of folds for cross-fitting}

\item{nboot}{A numeric indicating the number of desired bootstrap samples. If >0, uses bootstrap to obtain SEs. If =0, uses asymptotic analytical SEs.}

\item{data}{A data frame}
}
\value{
A list of results
}
\description{
drcmd is a function
}
