% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{drcmd_est}
\alias{drcmd_est}
\title{Obtain doubly-robust counterfactual mean estimates through cross-fitting}
\usage{
drcmd_est(
  Y,
  A,
  X,
  Z,
  R,
  m_learners,
  g_learners,
  r_learners,
  po_learners,
  eem_ind,
  Rprobs,
  k,
  c
)
}
\arguments{
\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{Z}{Dataframe containing all variables that are never subject to missingness}

\item{R}{Binary missingness indicator}

\item{m_learners}{A character vector containing learners to be used for the
outcome regression. User can specify 'hal' or a vector of SuperLearner libraries}

\item{g_learners}{A character vector containing learners to be used for the
propensity score. User can specify 'hal' or a vector of SuperLearner libraries}

\item{r_learners}{A character vector containing learners to be used for the
missingness indicator regression. User can specify 'hal' or a vector of
SuperLearner libraries}

\item{po_learners}{A character vector containing learners to be used for the
pseudo-outcome regression. User can specify 'hal' or a vector of SuperLearner libraries}

\item{eem_ind}{A logical indicating whether to use empirical efficiency maximization}

\item{Rprobs}{A vector of probabilities for R}

\item{k}{A numeric indicating the number of folds for cross-fitting}

\item{data}{A data frame}
}
\value{
A list of point estimates and standard errors for all estimands considered
}
\description{
Outer function for estimating counterfactual means through cross-fitting.
Calls \code{drcmd_est_fold} to obtain estimates for each cross-fitting fold
}
