% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_g}
\alias{est_g}
\title{Estimate the propensity score}
\usage{
est_g(idx, A, X, R, kappa_hat, g_learners)
}
\arguments{
\item{idx}{Indices to carry out estimation over}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{R}{Binary missingness indicator, where 0 indicates missing data}

\item{g_learners}{A character vector containing the names of the superlearner algorithms}

\item{Y}{Outcome variable. Can be continuous or binary}
}
\value{
A list containing the estimate of E\link{Y|A=a,X,W}
}
\description{
Function for obtaining propensity score estimates
}
