% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_kappa}
\alias{est_kappa}
\title{Estimate complete case propensity scores}
\usage{
est_kappa(idx, Z, R, r_learners)
}
\arguments{
\item{idx}{Indices to carry out estimation over}

\item{Z}{Dataframe containing all non-missing variables}

\item{R}{Binary missingness indicator, where 0 indicates missing data}

\item{r_learners}{A character vector specifying learners to be used for estimation}
}
\value{
A numeric vector containing the estimate of E\link{R|Z}
}
\description{
Function for obtaining complete case propensity score estimates
}
\examples{
\dontrun{
n <- 1000
Z <- rnorm(n)
R <- rbinom(n,1,plogis(Z))
Z <- data.frame(Z)
g_learners <- c('SL.glm','SL.gam')
est_g(idx=1:n, Z=Z, R=R, r_learners=r_learners)
}
}
