% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_m_a}
\alias{est_m_a}
\title{Function for estimating outcome regression}
\usage{
est_m_a(idx, Y, A, X, R, kappa_hat, m_learners)
}
\arguments{
\item{idx}{Indices to carry out estimation over}

\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{R}{Binary missingness indicator, where 0 indicates missing data}

\item{kappa_hat}{A numeric vector containing the fitted values of kappa}

\item{m_learners}{A character vector containing the names of the superlearner algorithms}
}
\value{
A list containing the estimate of E\link{Y|A=a,X}
}
\description{
Function for obtaining estimate of E\link{Y|A=a,X}
}
