% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{est_psi}
\alias{est_psi}
\title{Obtain estimates for current cross fitting fold}
\usage{
est_psi(idx, R, Z, kappa_hat, phi_hat, varphi_hat)
}
\arguments{
\item{idx}{A data frame}

\item{R}{A character string containing randomization variable name}

\item{Z}{A character vector containing the names of the variables in Z}

\item{kappa_hat}{A numeric vector containing the fitted values of kappa}

\item{phi_hat}{A list containing the estimate of E\link{phi|Z}}

\item{varphi_hat}{A list containing the estimate of E\link{phi|Z}}
}
\value{
A list of point estimates and standard errors for counterfactual means
and various counterfactual contrasts
}
\description{
Function for obtaining counterfactual mean estimates for the current
fold of the cross-fitting procedure
}
