% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{est_psi_tml}
\alias{est_psi_tml}
\title{Estimate causal effects through targeted maximum likelihood}
\usage{
est_psi_tml(
  idx,
  Y,
  A,
  X,
  R,
  Z,
  m_1_hat,
  m_0_hat,
  g_hat,
  kappa_hat,
  phi_1_hat,
  phi_0_hat,
  varphi_hat,
  maxits = 10,
  att = FALSE,
  atc = FALSE
)
}
\arguments{
\item{idx}{Indices of the training set}

\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Dataframe of covariates}

\item{R}{Complete case indicator vector}

\item{Z}{Dataframe containing all variables that are never subject to missingness}

\item{m_1_hat}{A numeric vector containing the fitted outcome predictions under A=1}

\item{m_0_hat}{A numeric vector containing the fitted outcome predictions under A=0}

\item{g_hat}{A numeric vector containing the fitted treatment propensity scores}

\item{kappa_hat}{A numeric vector containing the fitted values of kappa}

\item{phi_1_hat}{Estimates of EIC for E\link{Y(1)}}

\item{phi_0_hat}{Estimates of E\link{Y(0)}}

\item{varphi_hat}{A list containing estimate of E\link{phi_a|Z}}
}
\value{
A list containing updated estimates of the nuisance parameters
}
\description{
Function for obtaining estimates of causal estimands through
targeted maximum likelihood. To be called within drcmd_est_fold(). Updates
estimate of P(R=1|Z) and the plug-in estimator in a manner which removes the
plug-in bias incurred by an initial plug-in estimator
}
