% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{est_psi_tml}
\alias{est_psi_tml}
\title{Estimate causal effects through targeted maximum likelihood}
\usage{
est_psi_tml(
  idx,
  Y,
  A,
  X,
  R,
  Z,
  m_1_hat,
  m_0_hat,
  g_hat,
  kappa_hat,
  varphi_hat,
  maxits = 10
)
}
\description{
Function for obtaining estimates of causal estimands through
targeted maximum likelihood. To be called within drcmd_est_fold(). Updates
estimate of P(R=1|Z) and the plug-in estimator in a manner which removes the
plug-in bias incurred by an initial plug-in estimator
}
