% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{est_ses_crossfit}
\alias{est_ses_crossfit}
\title{Cross-fit SE estimation via influence curve contributions}
\usage{
est_ses_crossfit(res, y_bin)
}
\arguments{
\item{ics}{A list of influence curves for each fold, obtained within drcmd_est}
}
\value{
A list of cross-fit SEs
}
\description{
Estimates cross-fit SEs by taking contributions to influence curve
of each fold
}
