% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_varphi}
\alias{est_varphi}
\title{Perform pseudo-outcome regression with conventional loss function}
\usage{
est_varphi(idx, R, Z, phi_1_hat, phi_0_hat, po_learners, Y)
}
\arguments{
\item{idx}{Indices to carry out estimation over}

\item{R}{Binary missingness indicator, where 0 indicates missing data}

\item{Z}{Dataframe containing all non-missing variables}

\item{phi_1_hat}{A numeric vector containing the fitted values of phi under A=1}

\item{phi_0_hat}{A numeric vector containing the fitted values of phi under A=0}

\item{po_learners}{A character vector containing the names of the superlearner algorithms}
}
\value{
A list containing the estimate of E\link{phi|Z} for a=0 and a=1
}
\description{
Outer function for obtaining estimate of E\link{phi|Z}
}
