% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_varphi_eem}
\alias{est_varphi_eem}
\title{est_varphi_eem}
\usage{
est_varphi_eem(data, R, phi_1_hat, phi_0_hat, Rprobs, eem_ind, hal_ind, sl.lib)
}
\arguments{
\item{data}{A data frame}

\item{R}{A character string containing randomization variable name}

\item{phi_1_hat}{A list containing the estimate of phi under A=1}

\item{phi_0_hat}{A list containing the estimate of phi under A=0}

\item{eem_ind}{A logical value indicating whether to estimate via EEM}

\item{hal_ind}{A logical value indicating whether to estimate via HAL}

\item{sl.lib}{A character vector containing the names of the superlearner algorithms}
}
\value{
A list containing the estimate of E\link{phi|Z}
}
\description{
Function for obtaining estimate of E\link{phi|Z} via empirical efficiency maximization
}
