% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{est_varphi_main}
\alias{est_varphi_main}
\title{est_varphi}
\usage{
est_varphi_main(data, R, phi_1_hat, phi_0_hat, eem_ind, hal_ind, sl.lib)
}
\arguments{
\item{data}{A data frame}

\item{R}{A character string containing randomization variable name}

\item{eem_ind}{A logical value indicating whether to estimate via EEM}

\item{Y}{A character string containing outcome variable name}

\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing covariate variable names}

\item{Z}{A character vector containing the names of the variables in Z}

\item{phi_hat}{A list containing the estimate of phi}
}
\value{
A list containing the estimate of E\link{phi|Z}
}
\description{
Function for obtaining estimate of E\link{phi|Z}
}
