% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_missing_pattern}
\alias{find_missing_pattern}
\title{find_missing_pattern}
\usage{
find_missing_pattern(Y, A, X, W)
}
\arguments{
\item{Y}{A vector or data frame containing outcome values}

\item{A}{A vector or data frame  containing treatment variable values}

\item{X}{A data frame containing covariate values}

\item{W}{A data frame containing proxy variable values}

\item{data}{A data frame}

\item{R}{A vector containing missingness indicator variable}
}
\value{
A character string containing the missing pattern
}
\description{
Find the missing pattern in the data
}
