% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{get_nuisance_ests}
\alias{get_nuisance_ests}
\title{get_nuisance_ests}
\usage{
get_nuisance_ests(idx, Y, A, X, Z, R, hal_ind, sl.lib, Rprobs)
}
\arguments{
\item{Y}{A character string containing outcome variable name}

\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing covariate names}

\item{Z}{A character vector containing first stage variable names}

\item{R}{A character vector containing missing data indicator}

\item{hal_ind}{A logical indicating whether to use highly adaptive lasso}

\item{sl.lib}{A character string indicating the superlearner library to use}

\item{Rprobs}{A vector of probabilities for R}

\item{data}{A data frame}

\item{eem_ind}{A logical indicating whether to use ensemble of estimators}
}
\value{
A list of nuisance estimates
}
\description{
Function for obtaining estimates of all relevant nuisance functions
}
