% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{get_phi_hat}
\alias{get_phi_hat}
\title{get_phi_hat}
\usage{
get_phi_hat(Y, A, X, R, g_hat, m_a_hat, kappa_hat, hal_ind, sl.lib)
}
\arguments{
\item{Y}{A character string containing outcome variable name}

\item{A}{A character string containing treatment variable name}

\item{X}{A character vector containing covariate variable names}

\item{R}{A character string containing randomization variable name}

\item{g_hat}{A fitted propensity score model}

\item{m_a_hat}{A fitted outcome model}

\item{kappa_hat}{A fitted sampling probs model}

\item{hal_ind}{A logical indicating whether to use highly adaptive lasso}

\item{sl.lib}{A character string indicating the superlearner library to use}

\item{data}{A data frame}

\item{Z}{A character vector containing the names of the variables in Z}
}
\value{
A list containing the estimated EIFs for A=1 and A=0
}
\description{
Function for obtaining estimate of E\link{phi|Z}, where phi is the
full data EIF
}
