% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{get_phi_hat}
\alias{get_phi_hat}
\title{Contruct full data EIF estimate}
\usage{
get_phi_hat(Y, A, X, R, g_hat, m_a_hat, kappa_hat)
}
\arguments{
\item{Y}{Outcome variable. Can be continuous or binary}

\item{A}{A binary treatment variable (1=treated, 0=control)}

\item{X}{Dataframe containing baseline covariates}

\item{R}{Binary missingness indicator}

\item{g_hat}{Propensity score predictions}

\item{m_a_hat}{List of outcome predictions under A=0/1}

\item{kappa_hat}{Missingness probabilities}

\item{Z}{Dataframe containing all variables that are never subject to missingness}
}
\description{
Function for constructing estimate of the full data EIF, given
propensity score and outcome model fits
}
