% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.drcmd}
\alias{plot.drcmd}
\title{Plot results from drcmd object}
\usage{
\method{plot}{drcmd}(x, type = "All")
}
\description{
S3 method for plotting results from drcmd object. Plots are
available for the following: (1) psuedo outcome regression fit, (2) influence
curve distribution,
}
