% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.drcmd}
\alias{print.drcmd}
\title{Print drcmd object}
\usage{
\method{print}{drcmd}(x, ...)
}
\arguments{
\item{x}{An object of class drcmd}
}
\description{
S3 method for printing drcmd objects. Provides concise summary
of results, and prints values of optional arguments
}
