% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis.R
\name{SL.hal9001}
\alias{SL.hal9001}
\title{SuperLearner wrapper for the highly-adaptive lasso}
\usage{
SL.hal9001(Y, X, newX, family, obsWeights, ...)
}
\description{
Wrapper for the highly-adaptive lasso (HAL) algorithm
implemented in the hal9001 package
}
