% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.drcmd}
\alias{summary.drcmd}
\title{Summarize results from drcmd}
\usage{
\method{summary}{drcmd}(x, detail = FALSE, ...)
}
\arguments{
\item{detail}{Logical. If TRUE, print out values of user-supplied arguments}

\item{results}{An object of class drcmd}
}
\value{
No return value. Called for printing a detailed results summary
}
\description{
S3 method for summarizng drcmd results. Provides detailed summary
of estimation output, while also providing information on missingness
mechanism. Can optionally print out values of user-supplied arguments by setting
detail=TRUE
}
