% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcmd.R
\name{tml_updates}
\alias{tml_updates}
\title{Update nuisance functions with TML}
\usage{
tml_updates(
  idx,
  Y,
  A,
  X,
  R,
  Z,
  m_1_hat,
  m_0_hat,
  g_hat,
  kappa_hat,
  phi_1_hat,
  phi_0_hat,
  varphi_hat,
  maxits = 10
)
}
\arguments{
\item{idx}{Indices of the training set}

\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Dataframe of covariates}

\item{R}{Complete case indicator vector}

\item{Z}{Dataframe containing all variables that are never subject to missingness}

\item{m_1_hat}{A numeric vector containing the fitted outcome predictions under A=1}

\item{m_0_hat}{A numeric vector containing the fitted outcome predictions under A=0}

\item{g_hat}{A numeric vector containing the fitted treatment propensity scores}

\item{kappa_hat}{A numeric vector containing the fitted values of kappa}

\item{phi_1_hat}{Estimates of EIC for E\link{Y(1)}}

\item{phi_0_hat}{Estimates of E\link{Y(0)}}

\item{varphi_hat}{A list containing estimate of E\link{phi_a|Z}}
}
\value{
A list containing updated estimates of the nuisance parameters
}
\description{
Function for updating the nuisance functions m_a and kappa through
TML
}
