% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim}
\alias{trim}
\title{Trim vector (for numerical stability)}
\usage{
trim(x, val = .Machine$double.neg.eps)
}
\arguments{
\item{x}{A vector of values}

\item{val}{A small value to add to 0 and subtract from 1}
}
\value{
A vector with values trimmed to avoid numerical instability
}
\description{
Trims values of a vector to avoid numerical instability
}
\keyword{Internal}
